[!!!][TASK] Use a dedicated tube for each queue
Currently all queues share the same (default) tube but beanstalkd
supports multiple tubes.

This change uses a tube for each queue (with the tube name being
the queue name).

This is a breaking change if you expected the previous behavior of
beanstalkd to use the default tube for all interactions. In this
case adjust your code or simply name your queue "default".

Change-Id: I2aed01dc3e86dac6c05ac5061877359649f81a94